requirements.txt
adal==1.2.1
aiohttp==3.6.2
aiounittest==1.3.0
applicationinsights==0.11.9
async-timeout==3.0.1
attrs==19.3.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.25
botbuilder-ai==4.8.0
botbuilder-applicationinsights==4.8.0
botbuilder-core==4.8.0
botbuilder-schema==4.8.0
botframework-connector==4.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Django==2.2.6
djangorestframework==3.10.3
Flask==1.1.1
geographiclib==1.50
geopy==1.21.0
idna==2.8
idna-ssl==1.1.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpickle==1.2
MarkupSafe==1.1.1
msal==1.1.0
msrest==0.6.10
msrestazure==0.6.3
multidict==4.7.5
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
pycparser==2.20
PyJWT==1.5.3
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.8
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.4.2
